+++ libraries/AP_Math/vector3.h
@@ -175,10 +175,10 @@ public:
 
 };
 
-typedef Vector3<int>			Vector3i;
-typedef Vector3<unsigned int>	Vector3ui;
-typedef Vector3<long>			Vector3l;
-typedef Vector3<unsigned long>	Vector3ul;
+typedef Vector3<int16_t>		Vector3i;
+typedef Vector3<uint16_t>		Vector3ui;
+typedef Vector3<int32_t>		Vector3l;
+typedef Vector3<uint32_t>		Vector3ul;
 typedef Vector3<float>			Vector3f;
 
 #endif // VECTOR3_H
